Conflicts:
	src/engine/objects/GameMap.java

+++ src/engine/objects/MovableObject.java
@@ -3,7 +3,7 @@ package engine.objects;
 import com.badlogic.gdx.math.Vector2;
 
 public abstract class MovableObject extends GameObject {
-	public int gravity; 
+	public double gravity; 
 	public Vector2 velocity; 
 	boolean facingRight; 
 }
